Issue #4575: Modified NPathComplexityCheckTest.java and moved its input files to the npathcomplexity subdirectory
